# src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Swagger API",
    "description": "API desenvolvida ser apresentada na COP",
    "termsOfService": "",
    "contact": {
      "name": "DB",
      "url": "https://db.tec.br/",
      "email": "info@dbserver.com.br"
    },
    "license": {
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/auth": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Realizar Login de Usuário",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "email@admin.com",
                "senha": "1024"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retorna um Token de acesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "Token de acesso"
                    },
                    "tipo": {
                      "type": "string",
                      "example": "Bearer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Retorna um erro caso os dados estejam incorretos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Dados incorretos"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuario": {
      "post": {
        "tags": [
          "Usuário"
        ],
        "summary": "Cadastrar um Usuário",
        "operationId": "post",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "nome": "João",
                "email": "emailexemplo@email.com",
                "senha": "123456"
              }
            }
          },
          "required": true
        }
      },
      "get": {
        "tags": [
          "Usuário"
        ],
        "summary": "Busca todos os Usuários",
        "operationId": "get",
        "responses": {
          "200": {
            "description": "Retorna uma lista de Usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "example": {
                      "type": "object",
                      "$ref": "schemas/UsuarioSchema.json/exampleUsuario"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Retorna um erro caso os dados estejam incorretos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Dados incorretos"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Pessoa": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "João"
          },
          "idade": {
            "type": "integer",
            "format": "int32",
            "example": 19
          },
          "cpf": {
            "type": "string",
            "example": "860.111.980-82"
          }
        },
        "required": [
          "nome",
          "idade",
          "cpf"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "HTTP",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}